bot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-telegram-bot==21.7
httpx==0.28.1
pydantic==2.10.4
